Handle Rails 8 Result::IndexedColumn
In Rails 8, the first argument to instantiate is no longer a Hash. It is
a new, private class: ActiveRecord::Result::IndexedRow. This is similar
enough to Hash for Rails's purposes, but it doesn't respond to each.

This commit handles this in a lazy way by calling to_h on the first
argument to instantiate. This ensures it's in a format which the rest of
the method can handle, regardless of the Rails version.